se añadio el controlador de producto con el  respectivo router
