+++ src/app.js
@@ -1,10 +1,11 @@
 const express = require('express');
-
-const mainRouter = require('./routers/main')
-
 const app = express ();
 
-const controller = require('./controllers/paginas');
+const mainRouter = require('./routers/mainRouters')
+const productRouter = require('./routers/productRouters')
+
+app.set('view engine', 'ejs');
+app.set  ('views', './src/views');
 
 const port = process.env.PORT || 3000;
 app.listen(port,()=>console.log('server corriendo en el puerto'+port));
@@ -13,7 +14,7 @@ app.use(express.static('./src/Public'));
 
 
 app.use('/', mainRouter);
-app.use('/productCar', mainRouter);
-app.use('/productDar', mainRouter);
-app.use('/register', mainRouter);
-app.use('/login', mainRouter);
+app.use('/product', productRouter);
+
+
+
