+++ src/controllers/productController.js
@@ -1,0 +1,7 @@
+const controller = {
+    productDetail:(req,res)=>{
+        res.render('./productDetail.ejs');
+    },
+}
+
+module.exports = controller;
