+++ src/routers/mainRouters.js
@@ -1,13 +1,12 @@
 const express = require('express');
-
-const mainController = require('../controllers/paginas');
-
-
 const router = express.Router();
 
 
+const mainController = require('../controllers/mainController');
+
+
 router.get('/',mainController.home);
-router.get('/productDetail',mainController.productDetail);
+
 router.get('/productCar',mainController.productCar);
 router.get('/register',mainController.register);
 router.get('/login',mainController.login);
