+++ src/views/index.ejs
@@ -5,11 +5,11 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <script src="https://kit.fontawesome.com/2946b0c24d.js" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="./Estilos/home.css">
-    <link rel="stylesheet" href="./Estilos/productDetail.css">
-    <link rel="stylesheet" href="./Estilos/login.css">
-    <link rel="stylesheet" href="./Estilos/productCart.css">
-    <link rel="stylesheet" href="./Estilos/register.css">
+    <link rel="stylesheet" href="/Estilos/home.css">
+    <link rel="stylesheet" href="/Estilos/productDetail.css">
+    <link rel="stylesheet" href="/Estilos/login.css">
+    <link rel="stylesheet" href="/Estilos/productCart.css">
+    <link rel="stylesheet" href="/Estilos/register.css">
     <title>Home</title>
 </head>
 <body>
@@ -49,19 +49,19 @@
                     <img src="/images/item1reco.png" alt="item">
                     <h3>TITULO RELOJ</h3>
                     <P>Lorem ipsum dolor sit amet consectetur adipisicing elit. Perferendis aliquam quidem sit vel, dolores numquam suscipit possimus similique voluptatibus iure omnis reiciendis rerum voluptas ipsum? Mollitia hic magni esse laborum!</P>
-                    <a class='linkitem'href="/productDetail"><button>¡VER PRODUCTO!</button></a>
+                    <a class='linkitem'href="/product/detail"><button>¡VER PRODUCTO!</button></a>
                 </article>
                 <article class="sugerencia_item">
                     <img src="/images/item2reco.png" alt="item">
                     <h3>TITULO RELOJ</h3>
                     <P>Lorem ipsum dolor sit amet consectetur adipisicing elit. Perferendis aliquam quidem sit vel, dolores numquam suscipit possimus similique voluptatibus iure omnis reiciendis rerum voluptas ipsum? Mollitia hic magni esse laborum!</P>
-                    <a class='linkitem'href="/productDetail"><button>¡VER PRODUCTO!</button></a>
+                    <a class='linkitem'href="/product/detail"><button>¡VER PRODUCTO!</button></a>
                 </article>
                 <article class="sugerencia_item">
                     <img src="/images/item3reco.png" alt="item">
                     <h3>TITULO RELOJ</h3>
                     <P>Lorem ipsum dolor sit amet consectetur adipisicing elit. Perferendis aliquam quidem sit vel, dolores numquam suscipit possimus similique voluptatibus iure omnis reiciendis rerum voluptas ipsum? Mollitia hic magni esse laborum!</P>
-                    <a class='linkitem'href="/productDetail"><button>¡VER PRODUCTO!</button></a>
+                    <a class='linkitem'href="/product/detail"><button>¡VER PRODUCTO!</button></a>
                 </article>
             </div>
         </div>
