+++ src/views/productDetail.ejs
@@ -6,15 +6,15 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
     <script src="https://kit.fontawesome.com/2946b0c24d.js" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="./Estilos/home.css">
-    <link rel="stylesheet" href="./Estilos/productDetail.css">
-    <link rel="stylesheet" href="./Estilos/login.css">
-    <link rel="stylesheet" href="./Estilos/productCart.css">
-    <link rel="stylesheet" href="./Estilos/register.css">
+    <link rel="stylesheet" href="/Estilos/home.css">
+    <link rel="stylesheet" href="/Estilos/productDetail.css">
+    <link rel="stylesheet" href="/Estilos/login.css">
+    <link rel="stylesheet" href="/Estilos/productCart.css">
+    <link rel="stylesheet" href="/Estilos/register.css">
 </head>
 <body>
     <header>
-        <a class='linkmenus' href="/"><img class='logoMenu' src="./images/Logo_negro.png" alt="logo"></a>
+        <a class='linkmenus' href="/"><img class='logoMenu' src="/images/Logo_negro.png" alt="logo"></a>
         <form class="contenedorUno" >
             <div class="barraYlupa"><input class="search letra"  type="text"><i class="fas fa-search"></i></div>
             <div class="menus">
